Create deleteDuplicates2.py
Create a new dummy node and another reference to the dummy node. compare the first node with the next node value if they are equal move the head and also change your boolean status to true since you have found a duplicate. when the first while loop exists we check if we have found a duplicate just move the head pointer to the next element else attach the dummyNodes next value to the head and move the reference to the dummy node to the last position and make sure the next pointer is pointing to null